[user] Updated JSON generation for User and Session models
