+++ Sources/App/Models/Session/Session.generated.swift
@@ -18,28 +18,6 @@ extension Session {
   }
 }
 
-extension Session: Updateable {
-
-  public static var updateableKeys: [UpdateableKey<Session>] {
-    return [
-      UpdateableKey(Keys.userId, Identifier.self) { $0.userId = $1 },
-      UpdateableKey(Keys.token, String.self) { $0.token = $1 },
-      UpdateableKey(Keys.actual, Bool.self) { $0.actual = $1 }
-    ]
-  }
-}
-
-extension Session: JSONInitializable {
-
-  convenience init(json: JSON) throws {
-    self.init(
-      userId: try json.get(Keys.userId),
-      token: try json.get(Keys.token),
-      actual: try json.get(Keys.actual)
-    )
-  }
-}
-
 extension Session: Preparation {
 
   static func prepare(_ database: Database) throws {
@@ -53,18 +31,6 @@ extension Session: Preparation {
 
   static func revert(_ database: Database) throws {
     try database.delete(self)
-  }
-}
-
-extension Session: JSONRepresentable {
-
-  func makeJSON() throws -> JSON {
-    var json = JSON()
-    try json.set(Keys.id, id)
-    try json.set(Keys.userId, userId)
-    try json.set(Keys.token, token)
-    try json.set(Keys.actual, actual)
-    return json
   }
 }
